Restructured project into a more "Clean architecture" setup
