+++ src/Infrastructure/Trailblazor.Infrastructure/Usings.cs
@@ -1,0 +1,3 @@
+﻿global using Microsoft.EntityFrameworkCore;
+global using Microsoft.Extensions.Options;
+global using Duende.IdentityServer.EntityFramework.Options;
