+++ src/Server/Areas/Identity/Pages/Account/ConfirmEmail.cshtml.cs
@@ -6,12 +6,12 @@ using System;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Trailblazor.Infrastructure.Identity;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.WebUtilities;
-using Trailblazor.Server.Models;
 
 namespace Trailblazor.Server.Areas.Identity.Pages.Account
 {
