+++ src/Server/Areas/Identity/Pages/_ViewImports.cshtml
@@ -2,4 +2,4 @@
 @using Trailblazor.Server.Areas.Identity
 @using Trailblazor.Server.Areas.Identity.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@using Trailblazor.Server.Models
+@using Trailblazor.Infrastructure.Identity
