+++ src/Server/Pages/Shared/_LoginPartial.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Identity
-@using Trailblazor.Server.Models
+@using Trailblazor.Infrastructure.Identity
 
 @inject SignInManager<ApplicationUser> SignInManager
 @inject UserManager<ApplicationUser> UserManager
